requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==5.0.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.6
kombu==5.3.4
mutagen==1.47.0
prompt-toolkit==3.0.43
pycryptodomex==3.19.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
yt-dlp==2023.11.16
